Users & Authentication

+++ back_end/src/routes/EmpresaRoutes.py
@@ -11,48 +11,58 @@ main = Blueprint('empresa_blueprint', __name__)
 
 @main.route('/', methods=['GET'])
 def get_empresas():
-    # has_access = Security.verify_token(request.headers)
+    try:
+        empresas = EmpresaService.get_empresas()
+        if (len(empresas) > 0):
+            return jsonify([empresa.to_json() for empresa in empresas])
+        else:
+            return jsonify({'message': "NOTFOUND", 'success': True})
+    except CustomException:
+        return jsonify({'message': "ERROR", 'success': False})
 
-    # if has_access:
-        try:
-            empresas = EmpresaService.get_empresas()
-            if (len(empresas) > 0):
-                return jsonify([empresa.to_json() for empresa in empresas])
-            else:
-                return jsonify({'message': "NOTFOUND", 'success': True})
-        except CustomException:
-            return jsonify({'message': "ERROR", 'success': False})
-    # else:
-        # response = jsonify({'message': 'Unauthorized'})
-        # return response, 401
 @main.route('/', methods=['POST'])
 def post_empresa():
+    has_access = Security.verify_token(request.headers)
 
-    try:
-        nombre = request.json['nombre']
-        empresa = EmpresaService.post_empresa(nombre)
-        return jsonify(empresa.to_json()), 201
-    except CustomException:
-        return jsonify({'message': "ERROR", 'success': False})
+    if has_access:
+        try:
+            nombre = request.json['nombre']
+            empresa = EmpresaService.post_empresa(nombre)
+            return jsonify(empresa.to_json()), 201
+        except CustomException:
+            return jsonify({'message': "ERROR", 'success': False})
+    else:
+        response = jsonify({'message': 'Unauthorized'})
+        return response, 401
 
 @main.route('/<int:id_empresa>', methods=['PUT'])
-def put_proveedores(id_empresa: int):
+def put_empresa(id_empresa: int):
+    has_access = Security.verify_token(request.headers)
 
-    try:
-        if 'nombre' in request.json and request.json['nombre']:
-            nombre = request.json['nombre']
-            empresa = EmpresaService.put_empresa(nombre, id_empresa)
-            return jsonify(empresa.to_json()), 201
-        else:
-            return jsonify({'message': 'El nombre es requerido'}), 400
-    except CustomException:
-        return jsonify({'message': "ERROR", 'success': False})
+    if has_access:
+        try:
+            if 'nombre' in request.json and request.json['nombre']:
+                nombre = request.json['nombre']
+                empresa = EmpresaService.put_empresa(nombre, id_empresa)
+                return jsonify(empresa.to_json()), 201
+            else:
+                return jsonify({'message': 'El nombre es requerido'}), 400
+        except CustomException:
+            return jsonify({'message': "ERROR", 'success': False})
+    else:
+        response = jsonify({'message': 'Unauthorized'})
+        return response, 401
 
 @main.route('/<int:id_empresa>', methods = ['DELETE'])
 def delete_empresa(id_empresa: int):
+    has_access = Security.verify_token(request.headers)
 
-    try:
-        EmpresaService.delete_empresa(id_empresa)
-        return jsonify({'message': "Empresa borrada exitosamente"}), 200
-    except CustomException:
-        return jsonify({'message': "ERROR", 'success': False})
+    if has_access:
+        try:
+            EmpresaService.delete_empresa(id_empresa)
+            return jsonify({'message': "Empresa borrada exitosamente"}), 200
+        except CustomException:
+            return jsonify({'message': "ERROR", 'success': False})
+    else:
+        response = jsonify({'message': 'Unauthorized'})
+        return response, 401
